rename: thumbnail, badge 정리
