+++ apps/web/src/components/common/auction-form/attach-images-thumbnail.tsx
@@ -1,6 +1,6 @@
 import { X } from 'lucide-react';
 
-import Thumbnail from '../common/thumbnail';
+import Thumbnail from '@/components/ui/thumbnail';
 import { AttacedAuctionImageProps } from '@/lib/types/auction-prisma';
 
 const AttacedImagesThumbnail = ({ url, handleDelete }: AttacedAuctionImageProps) => {
