+++ apps/web/src/components/ui/thumbnail.tsx
